Add Cure action and rename Revive to Infect
The Revive "action" is renamed to "Infect" to have a nice balance with the "Cure" action. I did not want to use "Kill", "Decapitate", etc. as these felt too aggressive and gruesome, as opposed to playful, for action names when converting zombie code to the refactored code frame (Spirit). Also, revive can be used later for some other action.

git-svn-id: svn://localhost/Eulogy/trunk@24 312113c5-a7f5-e64c-8d66-78bef03f5f43