Add build specification for Bookmark Manager (Graveyard)
Note, the top-level VI must be named "BMark_MainWindow.vi" in order to load correctly as a custom bookmark manager. The Bookmark Manager library is reorganized to follow the file structure of the default bookmark manager. It was tricky to get all of the stuff to be placed in the appropriate locations to avoid loading errors. A project file is not needed. It also appears any file structure can be used, but for now the new structure is used to be consistent and hopefully avoid loading errors in the future.

git-svn-id: svn://localhost/Eulogy/trunk@65 312113c5-a7f5-e64c-8d66-78bef03f5f43