Fix VPIM build specification to use new Toolkit VI build specification
This avoids needing to include the all of the project folders in the VPIM build specification, or have the root of the project be the source for the VPIM build. Instead, the VPIM build only needs uses the `builds` directory and the project build specifications will properly build the various portions of code in the builds folder. The VPIM output is saved to the `packages` folder, which is also ignored by the source code.

git-svn-id: svn://localhost/Eulogy/trunk@71 312113c5-a7f5-e64c-8d66-78bef03f5f43