Add better shortcut menu plugin organization
A "Dig a Grave" menu item is added for any shortcut menu, unless it is from the front panel. A "Grave" submenu is added only if the affected object is a grave. The "Grave" submenu has the "Reanimate", "Smite", and "Bury" menu items. This groups the Eulogy functionality together and displays only menu items when appropriate.

This also yields a simpler codebase and better packaging.

git-svn-id: svn://localhost/Eulogy/trunk@86 312113c5-a7f5-e64c-8d66-78bef03f5f43