Add a Tools menu for the add-on
Currently, the Tools->Field R&D Services->Eulogy menu only provides a menu item to launch the Graveyard (Bookmark Manager). This is the same as using the View->Bookmark Manager but it avoids the popup dialog to select a manager. More menu items will probably be added when more features are added to the add-on.

git-svn-id: svn://localhost/Eulogy/trunk@91 312113c5-a7f5-e64c-8d66-78bef03f5f43