Change builds to prefix all VIs and Controls to avoid cross linking
The Source Distribution builds now include the Toolkit VIs if references and add prefixes to avoid naming collisions and cross linking. Some of the components of the add-on (Bookmark Manager and Quick Drop) could use the Toolkit VIs when installed, but this caused problems during development with conflicts when the Eulogy project was open and the Toolkit VIs were in memory. Specifically, if the Eulogy package was installed.

Also, if the developer/user changes the Toolkit VIs, which is easy to do, it will mess up the other components that depend on them. Thus, a "snapshot" of the toolkit VIs are taken when the components are built and then prefixed with a name to previous collisions and isolate each deployed component from each other and the toolkit VIs in the palette.

git-svn-id: svn://localhost/Eulogy/trunk@92 312113c5-a7f5-e64c-8d66-78bef03f5f43