Add automated build for saving to previous version of LabVIEW
Build VIs are created that will build each component following the project's build specifications, then it will save for previous each component to v2015. I still need to implement save for previous for LLBs, so the Shortcut Menu does not work yet.

Note, LabVIEW would crash if I tried to save for previous as a Post Build Action VI for each build specification. I tried the VI server as well. Nothing seemed work when save for previous was used in a Post Build Action for a project's build specification. The workaround was to create the various Build and Build All VIs and use the Application Builder API to build the project's build specification followed by a subVI that would save for previous. This also required changing the build specifications to output to an "output" directory and save for previous to the "builds" directory. This way, when the VIPM build is run, the v2015 is used instead of the v2017 without having to change the VPIM build specification.

If and when the VIPM Pro is purchased, the Build ALL VI can wrapped in a Pre-Build action for the VIPM Pro and then the process would be truly automated. Right now, it is a two step process. The Build All VI must be executed BEFORE the VPIM build specification is run.

git-svn-id: svn://localhost/Eulogy/trunk@93 312113c5-a7f5-e64c-8d66-78bef03f5f43