Refactor build automation
It was getting too difficult to save each component to a previous version. Instead, a Build VI is created that saves the entire project to a previous version (user selectable), then initiates the build for the specific version. The previous version project is saved to `builds\XX.X` folder, then the Build Previous Version Project VI is called via the command line using the appropriate version of LabVIEW. This VI runs automatically and builds the components. The components will be available in the `builds\XX.X\output` folder, where the `XX.X` is the LabVIEW version. The previous project can be opened in the correct version of LabVIEW to build the project, or the VIPB can be opened without having the start the older version labview but build the package for the older version. It is a lot easier to do this and eventually with the VIPM Pro, the build package file can be executed automatically as well.

This meant removing all of the individual build VIs because they are no longer needed. Now, to build the project, simply using the `Build.vi` file and select a version. Documentation still needs to be added for this build process.

git-svn-id: svn://localhost/Eulogy/trunk@94 312113c5-a7f5-e64c-8d66-78bef03f5f43