Add better automated build process
VIPM Pro has been purchased and the VIPM API has been installed. This made automating the building of the VI package easier and "one-click". Now, the "Build.vi" just needs to be run to build the VI package for LabVIEW 2015. Note, LabVIEW 2015 needs to be installed and this will require VIPM Pro to be available. The build can still happen without VIPM Pro, but it will have to be done manually. The instructions for manual building of the package are available in the new README and on the front panel of the "Build.vi".

git-svn-id: svn://localhost/Eulogy/trunk@95 312113c5-a7f5-e64c-8d66-78bef03f5f43