Refactor project and on-disk structure to reduce nested folders
There is no need for a `Public` and `Private` virtual/on-disk folders. The scope is indicated by glyphs in the project explorer, so of course we know that VIs in the `Private` virtual folder are of the private scope because they have the red key glyph. The private VIs are actually "support" VIs. So, all Private virtual folders are renamed to "Support" and the "Public" virtual folders and nesting are removed. Public VIs are the default for a project library. Scope-named virtual and on-disk folders just add unneeded implementation details and nesting that make navigation more difficult.

git-svn-id: svn://localhost/Eulogy/trunk@101 312113c5-a7f5-e64c-8d66-78bef03f5f43