Add examples to package and remove Is Grave from public API
I still need to decide on a coherent public API for finding all graves in a VI and project. Basically, I need to move some of the Bookmark Manager VIs into the Toolkit, but for each VI I do that I need to create tests and examples. The minimum that is needed is just the four action (dig, bury, reanimate, and smite) VI for the public API. The Is Grave.vi was only being used by the Shortcut Menu, so it was made to be a community scope instead of public and removed from the palette. It might be added back when more VIs are added, but for now it does not need to be included in the toolkit API.

git-svn-id: svn://localhost/Eulogy/trunk@122 312113c5-a7f5-e64c-8d66-78bef03f5f43