Fix highlighting not working
The %s format specifier used with the Scan from String function stops at the first whitespace character. This was truncating the paths to the VIs, which are used for partial identification of bookmarks to graves. A different format specifier, `%[^\n]`, was used, which will get all characters up to the first newline character (impossible for a path).

git-svn-id: svn://localhost/Eulogy/trunk@155 312113c5-a7f5-e64c-8d66-78bef03f5f43