Add dynamic shortcut menu generation
Depending on the location of the right-click within the bookmarks view, different shortcut menus will appear. For example, if the selection is for a VI, then the "All" variants of the reanimate, smite, and bury will appear, while a selection for a grave/zombie would show the single variants of these actions.

git-svn-id: svn://localhost/Eulogy/trunk@169 312113c5-a7f5-e64c-8d66-78bef03f5f43