Fix automated package build
The Unit Tests libraries were excluded from the Source Distribution builds for the Bookmark Manager and Toolkit, but since the Tests.lvlib were children of the component project libraries, the source distribution builds were missing dependencies. This was causing the VIP build to fail. The component project libraries were still looking for and contained a reference to the unit tests project libraries from the project even though these should not be included in the package/distribution.

The unit tests were moved from the component project libraries to the `tests` folder and Tests.lvlib. The `Support` virtual folders for each component were changed from Private to Community and the Tests.lvlib were made friends with the component's project library. This allowed unit tests of support VIs to function but still properly scoped sub-VIs.

git-svn-id: svn://localhost/Eulogy/trunk@210 312113c5-a7f5-e64c-8d66-78bef03f5f43