Change hook search path
The hook search path now appends the additional, code-defined paths instead of overriding the default system, user, and project paths. Before, if the additional search paths array was empty, then the default system, user, and project paths would be used, but if it was not empty, the system, user, and project paths would *not* even be included in the search path. It seems more reasonable to "fallback" to a default than to override.