Add Conditional Disable structures to each hook execute VI
If the `EULOGY_HOOKS_DISABLED` conditional variable is set to `TRUE`, then none of the hooks will execute; otherwise, all hooks will execute when an action is executed, i.e. Bury, Dig, Reanimate, and Smite. This provides users with a mechanism to turn off hooks for a specific project or temporarily. Some hooks could interfere with testing/debugging or could be slow.