Add conditional disable structures for each hook
A conditional disable structure is added to each hook that disables the hook based on a project condition variable specific to each hook:

- `EULOGY_POST_BURY_HOOK_DISABLED`
- `EULOGY_POST_DIG_HOOK_DISABLED`
- `EULOGY_POST_REANIMTE_HOOK_DISABLED`
- `EULOGY_POST_SMITE_HOOK_DISABLED`
- `EULOGY_PRE_BURY_HOOK_DISABLED`
- `EULOGY_PRE_DIG_HOOK_DISABLED`
- `EULOGY_PRE_REANIMATE_HOOK_DISABLED`
- `EULOGY_PRE_SMITE_HOOK_DISABLED`

If any of these symbols/variables are defined and set to `TRUE`, then the corresponding hook is disabled. Note, all hooks must be enabled, i.e. `EULOGY_HOOKS_DISABLED` is undefined or not `TRUE`.